refactor: quantity hook으로 분리
